Refactor: 리뷰 반영 ReservationApiController URI mapping  @RequsetMapping 으로 변경 컨트롤러 패키지 세분화 view, api
