+++ src/main/java/roomescape/controller/api/ReservationApiController.java
@@ -1,7 +1,8 @@
-package roomescape.controller;
+package roomescape.controller.api;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.dto.ReservationResponse;
 import roomescape.entity.Reservation;
@@ -9,11 +10,12 @@ import java.util.ArrayList;
 import java.util.List;
 
 @RestController
-public class ReservationController {
+@RequestMapping("/reservations")
+public class ReservationApiController {
 
     private List<Reservation> reservations = new ArrayList<>();
 
-    @GetMapping("/reservations")
+    @GetMapping
     public ResponseEntity<List<ReservationResponse>> getReservations() {
         List<ReservationResponse> reservationResponseList = reservations.stream()
                 .map(ReservationResponse::new)
